requirements.txt
snowflake-connector-python==3.4.0
streamlit==1.28.0
pandas==2.1.0
python-dotenv==1.0.0
